adds overview for internships (be-fr)

+++ backend/backend/backend/src/Domain/Entities/InternShip.cs
@@ -12,10 +12,10 @@ public class InternShip
     public Department Unit { get; set; }
     public int UnitId { get; set; }
     public string SchoolYear { get; set; }
-    public byte MaxStudents  { get; set; }
+    public int MaxStudents  { get; set; }
     public Location Location { get; set; }
     public int LocationId { get; set; }
-    public byte CurrentCountOfStudents { get; set; }
+    public int CurrentCountOfStudents { get; set; }
     public TrainingType RequiredTrainingType { get; set; }
     public IList<InternShipContentTranslation> Translations { get; set; }
 }
